Refine styles
