+++ src/Angor/Avalonia/AngorApp/Common/TransactionPreview/TransactionPreviewView.axaml
@@ -33,9 +33,9 @@
             <TextBlock>This is a preview of the transaction you will send.</TextBlock>
             <TextBlock>If everything is correct you can confirm it with the button below.</TextBlock>
         </StackPanel>
-        <HeaderedContentControl Theme="{StaticResource InlineHeaderedContentControl}" Header="Destination" Content="{Binding Destination.Name}" />
-        <HeaderedContentControl Theme="{StaticResource InlineHeaderedContentControl}" Header="Amount" Content="{Binding Destination.Amount, StringFormat={x:Static controls:MiscConverters.AmountBtcFormat}}" />
-        <HeaderedContentControl Theme="{StaticResource InlineHeaderedContentControl}" Header="Total Fee">
+        <HeaderedContentControl Theme="{StaticResource Inline}" Header="Destination" Content="{Binding Destination.Name}" />
+        <HeaderedContentControl Theme="{StaticResource Inline}" Header="Amount" Content="{Binding Destination.Amount, StringFormat={x:Static controls:MiscConverters.AmountBtcFormat}}" />
+        <HeaderedContentControl Theme="{StaticResource Inline}" Header="Total Fee">
             <HeaderedContentControl.Content>
                 <Panel HorizontalAlignment="Left">
                     <TextBlock VerticalAlignment="Center" Text="{Binding Feerate}" IsVisible="{Binding !CreateTransaction.IsExecuting^}" />
